spliting dashboard into multiple componenet
