+++ lib/screens/lobar_dashboard/dashboard_component.dart
@@ -1,0 +1,5 @@
+export 'dashboard_component/title_dashboard.dart';
+export 'dashboard_component/news_dashboard.dart';
+export 'dashboard_component/menu_dashboard.dart';
+export 'dashboard_component/agenda_bupati_dashboard.dart';
+export 'dashboard_component/banner_dashboard.dart';
