+++ lib/screens/lobar_dashboard/dashboard_component/title_dashboard.dart
@@ -1,0 +1,35 @@
+import 'package:flutter/material.dart';
+
+class TitleDashboard extends StatelessWidget {
+  const TitleDashboard({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      height: MediaQuery.of(context).size.height * 0.14,
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        mainAxisAlignment: MainAxisAlignment.start,
+        children: [
+          Container(
+            height: MediaQuery.of(context).size.height * 0.14,
+            decoration: const BoxDecoration(
+              boxShadow: [
+                BoxShadow(
+                  color: Colors.black38,
+                  offset: Offset(0, 2),
+                  blurRadius: 4,
+                  spreadRadius: 2,
+                )
+              ],
+              image: DecorationImage(
+                image: AssetImage('assets/images/header_img.png'),
+                fit: BoxFit.cover,
+              ),
+            ),
+          )
+        ],
+      ),
+    );
+  }
+}
